requirements.txt
streamlit==1.36.0
python-docx==1.1.2
scikit-learn==1.4.2
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.12.3
openai==1.35.6
python-dotenv==1.0.1
typer==0.12.3
rich==13.7.1
pypdf==5.1.0
